requirements.txt
# Oanda API
oandapyV20==0.7.2

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# データ分析
pandas==2.1.4
numpy==1.26.3
ta==0.11.0

# 機械学習
scikit-learn==1.4.0
tensorflow==2.15.0
joblib==1.3.2

# 深層学習・強化学習
torch==2.1.2
stable-baselines3==2.2.1
gymnasium==0.29.1

# 自然言語処理
transformers==4.36.2
sentencepiece==0.1.99

# ニュースAPI
newsapi-python==0.2.7
requests==2.31.0

# データ処理
beautifulsoup4==4.12.3
lxml==5.1.0

# データベース
sqlalchemy==2.0.25

# 環境変数管理
python-dotenv==1.0.0

# ログ
loguru==0.7.2

# スケジューリング
schedule==1.2.1

# テスト
pytest==7.4.4
pytest-asyncio==0.23.3
